remove unecessarys libs and adjust some points
