+++ Cargo.toml
@@ -13,7 +13,7 @@ homepage = "https://github.com/asimov-modules"
 repository = "https://github.com/asimov-modules/asimov-template-module"
 license = "Unlicense"
 keywords = ["asimov-module", "asimov", "ai"]
-categories = ["command-line-utilities", "text-processing"]              # TODO
+categories = ["command-line-utilities", "text-processing"]
 publish = false
 
 [features]
@@ -33,9 +33,7 @@ jq = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 anyhow = "1.0"
-reqwest = { version = "0.11", features = ["json"] }
 tracing-subscriber = "0.3"
-tokio = { version = "1.0", features = ["full"] }
 colored_json = "2.1"
 
 # Optional integrations:
@@ -46,6 +44,7 @@ clientele = { version = "0.3.8", default-features = false, features = [
     "clap",
     "std",
 ], optional = true }
+ureq = "3.1.0"
 
 [profile.release]
 opt-level = "z"
