+++ src/api/readwise.rs
@@ -1,21 +1,36 @@
 // This is free and unencumbered software released into the public domain.
 
-use crate::api::types::{BookListResponse, HighlightsResponse, ReadwiseConfig};
+use crate::api::types::{BookListResponse, HighlightsResponse};
 use anyhow::Result;
-use reqwest::Client;
+
+use ureq;
+
+#[derive(Debug, Clone)]
+pub struct ReadwiseConfig {
+    pub base_url: String,
+    pub access_token: String,
+}
+
+impl ReadwiseConfig {
+    pub fn new(access_token: String) -> Self {
+        Self {
+            base_url: "https://readwise.io/api/v2".to_string(),
+            access_token,
+        }
+    }
+
+    pub fn endpoint_url(&self, path: &str) -> String {
+        format!("{}{}", self.base_url, path)
+    }
+}
 
 pub struct ReadwiseClient {
     config: ReadwiseConfig,
-    client: Client,
 }
 
 impl ReadwiseClient {
     pub fn new(config: ReadwiseConfig) -> Result<Self> {
-        let client = Client::builder()
-            .timeout(std::time::Duration::from_secs(config.timeout))
-            .build()?;
-
-        Ok(Self { config, client })
+        Ok(Self { config })
     }
 
     fn auth_header(&self) -> String {
@@ -50,123 +65,120 @@ impl ReadwiseClient {
         url
     }
 
-    pub async fn fetch_highlights(
-        &self,
+    pub fn fetch_highlights(
+        &mut self,
         page_size: Option<usize>,
         page: Option<usize>,
     ) -> Result<HighlightsResponse> {
         let url = self.build_url_with_params("/highlights/", page_size, page);
-        let response = self
-            .client
-            .get(&url)
-            .header("Authorization", self.auth_header())
-            .send()
-            .await?;
 
-        let response_text = response.text().await?;
-        let highlights: HighlightsResponse = serde_json::from_str(&response_text)?;
-        Ok(highlights)
+        let mut response = ureq::get(&url)
+            .header("Authorization", &self.auth_header())
+            .call()
+            .map_err(|e| {
+                if e.to_string().contains("429") {
+                    anyhow::anyhow!("Rate limit exceeded (429). Please wait a minute before trying again. Consider using smaller page sizes to avoid hitting limits.")
+                } else {
+                    e.into()
+                }
+            })?;
+        let response_body: HighlightsResponse =
+            serde_json::from_str(&response.body_mut().read_to_string()?)?;
+        Ok(response_body)
     }
 
-    pub async fn fetch_booklist(
-        &self,
+    pub fn fetch_booklist(
+        &mut self,
         page_size: Option<usize>,
         page: Option<usize>,
     ) -> Result<BookListResponse> {
         let url = self.build_url_with_params("/books/", page_size, page);
-        let response = self
-            .client
-            .get(&url)
-            .header("Authorization", self.auth_header())
-            .send()
-            .await?;
 
-        let response_text = response.text().await?;
-        let booklist: BookListResponse = serde_json::from_str(&response_text)?;
-        Ok(booklist)
+        let mut response = ureq::get(&url)
+            .header("Authorization", &self.auth_header())
+            .call()
+            .map_err(|e| {
+                if e.to_string().contains("429") {
+                    anyhow::anyhow!("Rate limit exceeded (429). Please wait a minute before trying again. Consider using smaller page sizes to avoid hitting limits.")
+                } else {
+                    e.into()
+                }
+            })?;
+        let response_body: BookListResponse =
+            serde_json::from_str(&response.body_mut().read_to_string()?)?;
+        Ok(response_body)
     }
 
-    pub async fn fetch_highlight_tags(&self) -> Result<Vec<serde_json::Value>> {
+    pub fn fetch_highlight_tags(&mut self) -> Result<Vec<serde_json::Value>> {
         let mut all_tags = std::collections::HashMap::new();
         let mut page = 1;
         let page_size = 100;
 
         loop {
-            let highlights = self.fetch_highlights(Some(page_size), Some(page)).await?;
+            let highlights = self.fetch_highlights(Some(page_size), Some(page))
+            .map_err(|e| {
+                if e.to_string().contains("429") {
+                    anyhow::anyhow!("Rate limit exceeded while fetching highlights (429). Please wait a minute before trying again.")
+                } else {
+                    e
+                }
+            })?;
 
             if let Some(results) = highlights.results {
                 if results.is_empty() {
                     break;
                 }
 
+                let has_next = highlights.next.is_some();
+
                 for highlight in results {
                     if let Some(highlight_id) = highlight.id {
                         let tags_url =
                             self.endpoint_url(&format!("/highlights/{}/tags", highlight_id));
 
-                        let response = self
-                            .client
-                            .get(&tags_url)
-                            .header("Authorization", self.auth_header())
-                            .send()
-                            .await?;
+                        let mut response = ureq::get(&tags_url)
+                            .header("Authorization", &self.auth_header())
+                            .call()
+                            .map_err(|e| {
+                                if e.to_string().contains("429") {
+                                    anyhow::anyhow!("Rate limit exceeded while fetching tags (429). Please wait a minute before trying again.")
+                                } else {
+                                    e.into()
+                                }
+                            })?;
 
-                        let tags_response = response.text().await?;
+                        let response_body = response.body_mut().read_to_string()?;
 
-                        if let Ok(tags_data) =
-                            serde_json::from_str::<serde_json::Value>(&tags_response)
-                        {
-                            if let Some(tag_results) =
-                                tags_data.get("results").and_then(|r| r.as_array())
-                            {
-                                for tag in tag_results {
-                                    if let (Some(name), Some(id)) = (tag.get("name"), tag.get("id"))
-                                    {
-                                        if let (Some(name_str), Some(id_num)) =
-                                            (name.as_str(), id.as_u64())
-                                        {
-                                            all_tags.insert(
-                                                name_str.to_string(),
-                                                serde_json::json!({
-                                                    "name": name_str,
-                                                    "id": id_num
-                                                }),
-                                            );
-                                        }
-                                    }
+                        let tags_data: serde_json::Value = serde_json::from_str(&response_body)?;
+
+                        if let Some(tag_results) = tags_data["results"].as_array() {
+                            for tag in tag_results {
+                                if let (Some(name), Some(id)) =
+                                    (tag["name"].as_str(), tag["id"].as_u64())
+                                {
+                                    all_tags.insert(
+                                        name.to_string(),
+                                        serde_json::json!({
+                                            "name": name,
+                                            "id": id
+                                        }),
+                                    );
                                 }
                             }
                         }
                     }
                 }
 
-                page += 1;
+                if has_next {
+                    page += 1;
+                } else {
+                    break;
+                }
             } else {
                 break;
             }
         }
 
         Ok(all_tags.values().cloned().collect())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::api::types::types::ReadwiseConfig;
-
-    #[test]
-    fn test_readwise_config_new() {
-        let config = ReadwiseConfig::new("test_token".to_string());
-        assert_eq!(config.access_token, "test_token");
-        assert_eq!(config.base_url, "https://readwise.io/api/v2");
-    }
-
-    #[test]
-    fn test_auth_header() {
-        let config = ReadwiseConfig::new("test_token".to_string());
-        let client = ReadwiseClient::new(config).unwrap();
-        let header = client.auth_header();
-        assert_eq!(header, "Token test_token");
     }
 }
