--- src/api/types/mod.rs
@@ -1,2 +1,0 @@
-pub mod types;
-pub use types::{BookListResponse, HighlightsResponse, ReadwiseConfig};
