+++ src/fetcher/main.rs
@@ -1,21 +1,23 @@
 // This is free and unencumbered software released into the public domain.
-use asimov_readwise_module::providers::ReadwiseType;
+use asimov_readwise_module::api::types::ReadwiseType;
 use clap::Parser;
+use clientele::StandardOptions;
 
 #[derive(Parser)]
 #[command(name = "asimov-readwise-fetcher")]
 #[command(about = "URL protocol fetcher. Consumes a URL input, produces JSON output.")]
-struct Args {
+struct Options {
     #[arg(value_name = "INPUT-URL")]
-    input_url: String,
+    input_url: Option<String>,
 
-    /// Page size (number of items per page, default: fetch all)
     #[arg(long, value_name = "SIZE")]
     page_size: Option<usize>,
 
-    /// Specific page number to fetch (1-based, default: fetch all)
     #[arg(long, value_name = "NUM")]
     page: Option<usize>,
+
+    #[clap(flatten)]
+    flags: StandardOptions,
 }
 
 fn main() -> Result<clientele::SysexitsError, Box<dyn std::error::Error>> {
@@ -27,48 +29,56 @@ fn main() -> Result<clientele::SysexitsError, Box<dyn std::error::Error>> {
 
     clientele::dotenv().ok();
 
-    let args = Args::parse();
+    let options: Options = Options::parse();
+
+    if options.flags.version {
+        println!("{} {}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"));
+        return Ok(EX_OK);
+    }
+
+    if options.flags.license {
+        print!("{}", include_str!("../../UNLICENSE"));
+        return Ok(EX_OK);
+    }
+
+    let Some(input_url) = options.input_url else {
+        eprintln!("Missing INPUT-URL argument. Use --help for usage info");
+        return Ok(EX_USAGE);
+    };
 
     #[cfg(feature = "tracing")]
-    tracing_subscriber::fmt()
-        .with_writer(std::io::stderr)
-        .with_max_level(tracing_subscriber::filter::LevelFilter::WARN)
-        .init();
+    asimov_module::init_tracing_subscriber(&options.flags).expect("failed to initialize logging");
 
     let Some(api_key) = getenv::var_secret("READWISE_API_KEY") else {
         eprintln!("Missing READWISE_API_KEY. Run `asimov module config readwise`");
         return Ok(EX_CONFIG);
     };
 
-    let config = asimov_readwise_module::api::types::ReadwiseConfig::new(
+    let config = asimov_readwise_module::api::readwise::ReadwiseConfig::new(
         api_key.expose_secret().to_string(),
     );
 
-    let api = ReadwiseClient::new(config)?;
+    let mut api = ReadwiseClient::new(config)?;
 
-    let Some(provider) = find_provider_for(&args.input_url) else {
-        eprintln!("Unsupported URL: {}", args.input_url);
-        eprintln!("Make sure your URL starts with one of these patterns:");
-        eprintln!("  - https://readwise.io/highlights");
-        eprintln!("  - https://readwise.io/books");
-        eprintln!("  - https://readwise.io/tags");
+    let Some(provider) = find_provider_for(&input_url) else {
+        eprintln!(
+            "Unsupported URL: {}. Supported: highlights, books, tags",
+            input_url
+        );
         return Ok(EX_UNAVAILABLE);
     };
 
     let response = match provider.id {
         ReadwiseType::HIGHLIGHTS_ID => {
-            let rt = tokio::runtime::Runtime::new()?;
-            let highlights = rt.block_on(api.fetch_highlights(args.page_size, args.page))?;
+            let highlights = api.fetch_highlights(options.page_size, options.page)?;
             serde_json::to_string(&highlights)?
         },
         ReadwiseType::BOOKLIST_ID => {
-            let rt = tokio::runtime::Runtime::new()?;
-            let booklist = rt.block_on(api.fetch_booklist(args.page_size, args.page))?;
+            let booklist = api.fetch_booklist(options.page_size, options.page)?;
             serde_json::to_string(&booklist)?
         },
         ReadwiseType::TAGS_ID => {
-            let rt = tokio::runtime::Runtime::new()?;
-            let tags = rt.block_on(api.fetch_highlight_tags())?;
+            let tags = api.fetch_highlight_tags()?;
             serde_json::to_string(&tags)?
         },
         _ => {
@@ -86,61 +96,4 @@ fn main() -> Result<clientele::SysexitsError, Box<dyn std::error::Error>> {
     }
 
     Ok(EX_OK)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use asimov_readwise_module::find_provider_for;
-    use asimov_readwise_module::providers::ReadwiseType;
-
-    #[test]
-    fn test_find_provider_for_highlights() {
-        let url = "https://readwise.io/highlights";
-        let provider = find_provider_for(url);
-        assert!(provider.is_some());
-        assert_eq!(provider.unwrap().id, ReadwiseType::HIGHLIGHTS_ID);
-    }
-
-    #[test]
-    fn test_find_provider_for_books() {
-        let url = "https://readwise.io/books";
-        let provider = find_provider_for(url);
-        assert!(provider.is_some());
-        assert_eq!(provider.unwrap().id, ReadwiseType::BOOKLIST_ID);
-    }
-
-    #[test]
-    fn test_find_provider_for_tags() {
-        let url = "https://readwise.io/tags";
-        let provider = find_provider_for(url);
-        assert!(provider.is_some());
-        assert_eq!(provider.unwrap().id, ReadwiseType::TAGS_ID);
-    }
-
-    #[test]
-    fn test_find_provider_for_unsupported_url() {
-        let url = "https://example.com/api/books";
-        let provider = find_provider_for(url);
-        assert!(provider.is_none());
-    }
-
-    #[test]
-    fn test_provider_brand() {
-        let url = "https://readwise.io/highlights";
-        let provider = find_provider_for(url).unwrap();
-        assert_eq!(provider.brand, "Readwise");
-    }
-
-    #[test]
-    fn test_args_with_pagination() {
-        let args = Args {
-            input_url: "https://readwise.io/highlights".to_string(),
-            page_size: Some(100),
-            page: Some(2),
-        };
-        assert_eq!(args.input_url, "https://readwise.io/highlights");
-        assert_eq!(args.page_size, Some(100));
-        assert_eq!(args.page, Some(2));
-    }
 }
