+++ src/providers/mod.rs
@@ -1,6 +1,4 @@
 pub mod provider;
 pub mod readwise;
-pub mod types;
 
 pub use provider::*;
-pub use types::*;
